safe RemoteUpdateFile de-serialization if one of trails' classes not found in runtime

+++ bloodseeker/src/main/java/com/am1goo/bloodseeker/update/RemoteUpdateReader.java
@@ -33,7 +33,7 @@ public class RemoteUpdateReader extends DataInputStream {
             throw new IOException("wrong header");
     }
 
-    public int readVersion() throws IOException {
+    public short readVersion() throws IOException {
         return readShort();
     }
